refactor company stat dto to take overtime dto
closes #6392
